refactor(graph): extract the tag queries out of queries.ts (#1838)
`queries.ts` is the largest module in main at 1,294 lines, carrying six
unrelated query families behind section comments — the shape `indexers.ts` had
before #1624 split it. This starts the same split with the most self-contained
family, and establishes `graph/queries/` as its home.

Pure move: five functions, no signature changes, no logic touched. `queries.ts`
re-exports them, so `graph/index` and every `import … from './queries'` caller
is unchanged — the facade shape `indexers.ts` uses for `./indexers/*`. The
whole suite passes untouched, which is the point: a behaviour-preserving move
should need no test edits to prove it.

One correction to the plan in the issue, which named lines 348-497 as "six
exported functions, no cross-family calls". `listAllSources` sits inside that
range and is neither: it's a source query, and it calls `collectSourceMetadata`
from the source-detail family. It stays put and moves with source-detail. The
five that moved are genuinely tag-keyed and call nothing outside `./state`.

Remaining families, for the follow-ups: source detail (~460 lines, nine private
helpers), links (~325), aliases/frontmatter (~100), citations/anchors (~90),
SPARQL (~150).
